requirements.txt
python-binance==1.0.12
requests==2.25.1
requests-oauthlib==1.3.0